+++ packages/expo-sensors/ios/PedometerModule.swift
@@ -83,6 +83,8 @@ public final class PedometerModule: Module {
 
     OnStopObserving {
       stopUpdates()
+      watchStartDate = nil
+      watchHandler = nil
     }
 
     OnAppEntersBackground {
@@ -109,8 +111,6 @@ public final class PedometerModule: Module {
     }
 
     pedometer.stopUpdates()
-    watchStartDate = nil
-    watchHandler = nil
   }
 }